CPS-655: Fix new site deployments
