+++ CRM/Civicase/Upgrader/Steps/Step0016.php
@@ -1,6 +1,6 @@
 <?php
 
-use CRM_Civicase_Service_CaseCategoryCustomFieldsSetting as CaseCategoryCustomFieldsSetting;
+use CRM_Civicase_Setup_AddSingularLabels as AddSingularLabels;
 
 /**
  * Adds singular labels to case type categories.
@@ -14,33 +14,9 @@ class CRM_Civicase_Upgrader_Steps_Step0016 {
    *   True when the upgrader runs successfully.
    */
   public function apply() {
-    $this->addSingularLabelToCaseCategories();
+    (new AddSingularLabels())->apply();
 
     return TRUE;
-  }
-
-  /**
-   * Adds the singular label value for each case category.
-   *
-   * If the case category ends in an S, it will remove it.
-   */
-  private function addSingularLabelToCaseCategories() {
-    $caseCategoryCustomFields = new CaseCategoryCustomFieldsSetting();
-    $caseTypeCategories = civicrm_api3('OptionValue', 'get', [
-      'sequential' => '1',
-      'option_group_id' => 'case_type_categories',
-    ]);
-
-    foreach ($caseTypeCategories['values'] as $caseTypeCategory) {
-      $isLabelLastCharacterS = substr(strtolower($caseTypeCategory['label']), -1) === 's';
-      $singularLabel = $isLabelLastCharacterS
-        ? substr($caseTypeCategory['label'], 0, -1)
-        : $caseTypeCategory['label'];
-
-      $caseCategoryCustomFields->save($caseTypeCategory['value'], [
-        'singular_label' => $singularLabel,
-      ]);
-    }
   }
 
 }
